boldreams/requirements.txt
numpy~=1.22.4
igl~=2.2.1
scipy~=1.8.0
h5py~=3.2.1
torch~=1.12.1
tqdm~=4.65.0
pillow~=9.1.1
torchvision~=0.13.1
clip~=1.0
lucent~=0.1.0
torchextractor~=0.3.0